Update to VPAY01 schema
Per the DBAs:
This is a standard we are now following.
Schemas are security containers and when all objects are consolidated in
one schema, we lose the ability to granulate and isolate security
policies.

Since we are going to become more integrated with Optum, that was a
motivator in going in that direction.